healthsure/requirements.txt
asgiref==3.6.0
Django==4.1.7
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
node==1.2
npm==0.1.1
odict==1.9.0
optional-django==0.1.0
plumber==1.7
sqlparse==0.4.3
tzdata==2022.7
zope.component==5.1.0
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.6
zope.hookable==5.4
zope.interface==5.5.2
zope.lifecycleevent==4.4
zope.proxy==5.0.0
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
node==1.2
npm==0.1.1
odict==1.9.0
optional-django==0.1.0
plumber==1.7
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2022.7
zope.component==5.1.0
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.6
zope.hookable==5.4
zope.interface==5.5.2
zope.lifecycleevent==4.4
zope.proxy==5.0.0
